Añadida opción borrar usuario y proyectos restringida. Cambiado estilo para comprobar en navegadores y máquinas no Firefox.

+++ src/protected/pages/Project/ProjectList.php
@@ -37,51 +37,71 @@ class ProjectList extends TPage
 		$this->list->EditItemIndex = -1;
 		$this->showList();
 	}
-
-	public function deleteEntryItem($sender, $param)
-	{
-		$id = $this->list->DataKeys[$param->Item->ItemIndex];
-		$this->getProjectDao()->deleteProject($id);
-		$this->refreshEntryList();
-	}
 			
 	public function updateEntryItem($sender, $param)
 	{		
-		if(!$this->Page->IsValid)
-			return;
-			
 		$item = $param->Item;
 		$id = $this->list->DataKeys[$item->ItemIndex];
-
-		// configure project changes
+		$project = $this->getProjectDao()->getProjectByID($id);
+		$newManager = $item->manager->SelectedValue;
+		
+		// check project manager changes
+		if ($project->ManagerID != $newManager){
+			$this->getWorkerDao()->updateParticipation(
+				$project->ManagerID, $project->Title, 'Analista', 50);
+			if ($this->getWorkerDao()->participationExists($newManager, $project->Title))
+				$this->getWorkerDao()->updateParticipation(
+					$newManager, $project->Title, 'Jefe de proyecto', 50);
+			else
+				$this->getWorkerDao()->addParticipation(
+					$newManager, $project->Title, 'Jefe de proyecto', 50);
+		}
+		
+		// configure new project with changes
 		$project = new ProjectRecord;
 		$project->Title = $id;
-		$project->ManagerID = $param->Item->manager->SelectedValue;
-		$project->Description = $param->Item->description->Text;
-		$project->Burchet = floatval($param->Item->burchet->Text);	
+		$project->ManagerID = $item->manager->SelectedValue;
+		$project->Description = $item->description->Text;
+		$project->Burchet = floatval($item->burchet->Text);	
 		$this->getProjectDao()->updateProject($project);
 					
 		// update table
 		$this->refreshEntryList();
 	}
+	
+	public function deleteEntryItem($sender, $param)
+	{
+		$id = $this->list->DataKeys[$param->Item->ItemIndex];
+		
+		// check if deletion is possible
+		if (!$this->getProjectDao()->projectWorkersExists($id)){
+			$manager = $this->getProjectDao()->getProjectManager($id);
+			$this->getWorkerDao()->deleteParticipation($manager, $id);
+			$this->getProjectDao()->deleteProject($id);
+		} else
+			$this->views->ActiveViewIndex = 1;
+
+		$this->refreshEntryList($sender, $param);
+	}
 
 	public function EntryItemCreated($sender, $param)
 	{
+		$item = $param->Item;
 		if($param->Item->ItemType == 'EditItem' && $param->Item->DataItem)
 		{
-			$param->Item->manager->DataSource =	
-				$this->getManagers($param->Item->DataItem->ManagerID);	
-			$param->Item->manager->dataBind();
-			$param->Item->manager->SelectedValue = $param->Item->DataItem->ManagerID;
+			$item->manager->DataSource = $this->getManagers($item->DataItem->ManagerID);	
+			$item->manager->dataBind();
+			$item->manager->SelectedValue = $param->Item->DataItem->ManagerID;
 		}
 	}
 	
-	public function getManagers($manager){
+	public function getManagers($manager)
+	{
 		$managers = array();
-		foreach ($this->getWorkerDao()->getProjectManagers() as $man) 
-			$managers[$man] = $man;
+		foreach ($this->getWorkerDao()->getProjectManagers() as $man)
+			$managers[$man['UserID']] = $man['Worker'];
 		// add actual manager
-		$managers[$manager] = $manager;
+		$managers[$manager] = $this->getWorkerDao()->getWorkerName($manager);
 		return $managers;
 	}
 }
